Move appConfig outside of the public folder
